chore(llvm): Create a new scope for each function
