+++ src/backend_llvm/src/transformer/mod.rs
@@ -1,47 +1,16 @@
-use std::{cell::RefCell, collections::HashMap, rc::Rc};
+use std::rc::Rc;
 
 use inkwell as llvm;
 use swc_ecma_ast as ast;
 
 use crate::{types::CompiledExpression, BackendError};
 
-#[derive(Debug)]
-pub struct Scope<'ctx> {
-    pub variables: RefCell<HashMap<String, CompiledExpression<'ctx>>>,
-}
-
-impl<'ctx> Scope<'ctx> {
-    pub fn new() -> Self {
-        Self {
-            variables: RefCell::new(HashMap::new()),
-        }
-    }
-
-    pub fn define_variable(&self, identifier: String, reference: CompiledExpression<'ctx>) {
-        let mut variables = self.variables.borrow_mut();
-        if variables.contains_key(&identifier) {
-            panic!("Unable to define variable twice");
-        } else {
-            variables.insert(identifier, reference);
-        }
-    }
-
-    pub fn get_variable(
-        &self,
-        identifier: String,
-    ) -> Result<CompiledExpression<'ctx>, BackendError> {
-        if let Some(reference) = self.variables.borrow().get(&identifier) {
-            Ok(reference.clone())
-        } else {
-            panic!("Unknown variable: {}", identifier);
-        }
-    }
-}
+use self::scope::{Scope, ScopeStack};
 
 #[derive(Debug)]
 pub struct Transformer<'ctx> {
     // Our context
-    scope: Rc<Scope<'ctx>>,
+    scope_stack: ScopeStack<'ctx>,
     // LLVM context
     context: &'ctx llvm::context::Context,
     module: llvm::module::Module<'ctx>,
@@ -52,7 +21,7 @@ impl<'ctx> Transformer<'ctx> {
     pub fn new(context: &'ctx llvm::context::Context) -> Self {
         Transformer {
             // our
-            scope: Rc::new(Scope::new()),
+            scope_stack: ScopeStack::new(),
             // llvm
             context,
             module: context.create_module("module"),
@@ -62,7 +31,9 @@ impl<'ctx> Transformer<'ctx> {
 
     /// Transform module
     pub fn transform_fn(&mut self, fn_declr: ast::FnDecl) -> Result<(), BackendError> {
+        self.scope_stack.push(Rc::new(Scope::new()));
         self.compile_fn(fn_declr)?;
+        self.scope_stack.pop();
 
         Ok(())
     }
@@ -84,7 +55,9 @@ impl<'ctx> Transformer<'ctx> {
             if let Some(init) = d.init {
                 let default = self.compile_expr(init)?;
 
-                self.scope.define_variable(identifier, default);
+                self.scope_stack
+                    .scope()
+                    .define_variable(identifier, default);
             };
         }
 
@@ -95,7 +68,7 @@ impl<'ctx> Transformer<'ctx> {
         &mut self,
         ident: ast::Ident,
     ) -> Result<CompiledExpression<'ctx>, BackendError> {
-        self.scope.get_variable(ident.sym.to_string())
+        self.scope_stack.scope().get_variable(ident.sym.to_string())
     }
 
     fn compile_expr(
@@ -155,4 +128,5 @@ impl<'ctx> Transformer<'ctx> {
 
 mod binary;
 mod function;
+mod scope;
 mod statement;
